# docs/api-contract.yaml
openapi: 3.1.0
info:
  title: Ride-hailing API (Baghdad v1)
  version: 1.0.0
  description: |
    Binding API contract. CLAUDE.md §12.1: no endpoint may exist in the code that
    is not in this file. Adding an endpoint means changing this file first, in
    its own commit, with a reason.

    Conventions:
      * All money fields are named `*_iqd`, are integers, and are whole Iraqi
        Dinars (CLAUDE.md §6.1). A fractional value in any `*_iqd` field is a
        contract violation, not a rounding detail.
      * All timestamps are RFC 3339 UTC. Clients render in Asia/Baghdad.
      * All errors are RFC 9457 `application/problem+json` (CLAUDE.md §9).
      * Every request carries `X-Request-Id`; the server generates one if absent
        and always echoes it.
      * Bearer JWT auth unless the operation is explicitly marked `security: []`.
servers:
  - url: /v1
...
security:
  - bearerAuth: []
paths:
...
  /driver/availability:
    put:
      tags: [driver]
      operationId: setDriverAvailability
      summary: |
        Go online or offline. Going ONLINE requires a current position so the
        driver can be placed in the Redis geo set immediately (CLAUDE.md §3.1);
        without it the driver would be "online" but unmatchable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [availability]
              properties:
                availability:
                  type: string
                  enum: [ONLINE, OFFLINE]
                position: { $ref: '#/components/schemas/LatLng' }
      responses:
        '200':
          description: Applied.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverState' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Cannot go OFFLINE while assigned to a live ride.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          $ref: '#/components/responses/ValidationError'
  /driver/location:
    post:
      tags: [driver]
      operationId: reportDriverLocation
      summary: |
        Report one or more position samples. Accepts a BATCH because the driver
        app buffers while offline and flushes on reconnect (CLAUDE.md §5.3).
        Writes to Redis only; Postgres is written by the 30s flush job
        (CLAUDE.md §3.1). Returns 202 — the caller must not wait on durability.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [samples]
              properties:
                samples:
                  type: array
                  minItems: 1
                  maxItems: 200
                  items: { $ref: '#/components/schemas/LocationSample' }
      responses:
        '202':
          description: Accepted for processing.
          content:
            application/json:
              schema:
                type: object
                required: [accepted]
                properties:
                  accepted: { type: integer, minimum: 0 }
                  rejected: { type: integer, minimum: 0 }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /driver/offers/current:
    get:
      tags: [driver]
      operationId: getCurrentOffer
      summary: |
        The offer outstanding for this driver, if any. Polling fallback for when
        the push/WebSocket path failed — the driver app must not depend solely
        on FCM delivery.
      responses:
        '200':
          description: An offer is outstanding.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RideOffer' }
        '204':
          description: No outstanding offer.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /rides/{rideId}/accept:
    parameters:
      - $ref: '#/components/parameters/RideId'
    post:
      tags: [driver]
      operationId: acceptRide
      summary: |
        Accept an offer. CLAUDE.md §5.1: the winner is decided by
        `SET ride:{rideId}:claim {driverId} NX PX 30000`. Exactly one caller can
        get 200; every other concurrent caller gets 409 with
        `type: .../ride-already-claimed`.
      responses:
        '200':
          description: This driver won the claim and is now assigned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ride' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Another driver already claimed it, or the offer expired.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /rides/{rideId}/decline:
    parameters:
      - $ref: '#/components/parameters/RideId'
    post:
      tags: [driver]
      operationId: declineRide
      responses:
        '204':
          description: Declined; the ride is re-offered to the next candidate.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/InvalidTransition'
  /rides/{rideId}/arrived:
    parameters:
      - $ref: '#/components/parameters/RideId'
    post:
      tags: [driver]
      operationId: markDriverArrived
      responses:
        '200':
          description: ACCEPTED → DRIVER_ARRIVED.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ride' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
  /rides/{rideId}/start:
    parameters:
      - $ref: '#/components/parameters/RideId'
    post:
      tags: [driver]
      operationId: startRide
      responses:
        '200':
          description: DRIVER_ARRIVED → IN_PROGRESS.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ride' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
  /rides/{rideId}/complete:
    parameters:
      - $ref: '#/components/parameters/RideId'
    post:
      tags: [driver]
      operationId: completeRide
      summary: |
        Finish the trip. Settles the fare, writes the `payments` row and the
        balanced ledger entries in ONE database transaction with the status
        change. Idempotent by ride: completing an already-COMPLETED ride returns
        200 with the same body and writes nothing.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                actualDistanceM:
                  type: integer
                  minimum: 0
                  maximum: 1000000
                  description: Odometer distance from the driver app. Omitted → the estimate is used.
      responses:
        '200':
          description: Completed and settled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RideSettlement' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ---------------------------------------------------------------------------
  # Fare negotiation
  #
  # Owner decision 2026-08-25, recorded in CLAUDE.md §2 SCOPE EXPANSION.
  #
  # This does NOT replace nearest-driver dispatch. §5.1's atomic claim still
  # decides who gets a ride, and direct dispatch stays as the fallback when
  # nobody bids. Negotiation adds a window, before dispatch, in which drivers
  # commit to a price.
  #
  # The load-bearing idea: a driver's bid IS their acceptance. It is a binding
  # commitment at that fare. When the rider selects one, the server takes the
  # Redis claim on that driver's behalf and runs
  # REQUESTED -> OFFERED -> ACCEPTED through RideStateMachine unchanged. No new
  # transition rule, no bypass of the claim, and every existing guarantee about
  # one-winner still holds by construction rather than by a second mechanism.
  #
  # Disabled by default: platform_config.negotiation_enabled is seeded 'false'.
  # With it false these endpoints return 404 and ride creation behaves exactly
  # as it does today.
  # ---------------------------------------------------------------------------
...
  /driver/wallet:
    get:
      tags: [driver]
      operationId: getMyWallet
      summary: Balance derived from the ledger (CLAUDE.md §6.4), never a stored counter.
      responses:
        '200':
          description: Wallet.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wallet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /driver/wallet/entries:
    get:
      tags: [driver]
      operationId: listMyWalletEntries
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
        - name: cursor
          in: query
          schema: { type: string, maxLength: 200 }
      responses:
        '200':
          description: Page of the caller's own ledger entries, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LedgerEntryPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /driver/subscription/plans:
    get:
      tags: [driver]
      operationId: listSubscriptionPlans
      summary: |
        The plans a driver may buy. Active plans only, cheapest first.

        Readable whether or not `subscription_required` is on: a driver should
        be able to see what a subscription costs before it gates anything, and
        the same list is what the admin grant path charges against.
      responses:
        '200':
          description: Active plans.
          content:
            application/json:
              schema:
                type: object
                required: [plans]
                properties:
                  plans:
                    type: array
                    items: { $ref: '#/components/schemas/SubscriptionPlan' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /driver/subscription:
    get:
      tags: [driver]
      operationId: getMySubscription
      summary: |
        The caller's live subscription, or null.

        `expiresAt` is the authoritative date and is compared against the server
        clock, not against the stored status - a status column is only as fresh
        as the last sweep, and between two sweeps it would say ACTIVE for a
        lapsed subscription.
      responses:
        '200':
          description: The caller's subscription, or null when they have none.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverSubscriptionOrNull' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RideId:
      name: rideId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Missing or invalid token.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not permitted.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Not found, or not visible to this caller.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InvalidTransition:
      description: The requested ride state transition is not permitted from the current state.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.rideapp.iq/problems/invalid-ride-transition
            title: Invalid ride transition
            status: 409
            detail: Cannot move ride from COMPLETED to IN_PROGRESS.
            fromState: COMPLETED
            toState: IN_PROGRESS
    ValidationError:
      description: Request failed schema validation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [type, title, status]
      properties:
        type:     { type: string, format: uri }
        title:    { type: string }
        status:   { type: integer }
        detail:   { type: string }
        instance: { type: string }
        requestId: { type: string }
        errors:
          type: array
          description: Field-level validation failures.
          items:
            type: object
            required: [path, message]
            properties:
              path:    { type: string }
              message: { type: string }
      additionalProperties: true
    LatLng:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number, minimum: -90,  maximum: 90 }
        lng: { type: number, minimum: -180, maximum: 180 }
    LocationSample:
      type: object
      required: [lat, lng, recordedAt]
      properties:
        lat:         { type: number, minimum: -90,  maximum: 90 }
        lng:         { type: number, minimum: -180, maximum: 180 }
        accuracyM:   { type: number, minimum: 0, maximum: 100000 }
        headingDeg:  { type: number, minimum: 0, maximum: 360 }
        speedMps:    { type: number, minimum: 0, maximum: 200 }
        recordedAt:  { type: string, format: date-time }
    RideStatus:
      type: string
      enum:
        - REQUESTED
        - OFFERED
        - ACCEPTED
        - DRIVER_ARRIVED
        - IN_PROGRESS
        - COMPLETED
        - CANCELLED_BY_RIDER
        - CANCELLED_BY_DRIVER
        - CANCELLED_IN_TRIP
        - EXPIRED
        - NO_DRIVERS_FOUND
...
    PublicUser:
      type: object
      description: |
        How a counterparty is shown to the other side of a ride. Deliberately
        has NO phone field. ACCEPTANCE_CHECKLIST.md check 5 requires that a
        driver cannot see the rider's number after the trip; the way this is
        guaranteed is that the field does not exist in the response shape at all.
      required: [id, displayName]
      properties:
        id:          { type: string, format: uuid }
        displayName: { type: string }
        rating:      { type: number, nullable: true, minimum: 1, maximum: 5 }
        vehicle:
          type: object
          nullable: true
          properties:
            plate: { type: string }
            model: { type: string }
            color: { type: string }
    DriverState:
      type: object
      required: [availability, isSuspended]
      properties:
        availability: { type: string, enum: [OFFLINE, ONLINE, ON_TRIP] }
        isSuspended:  { type: boolean }
        vehicle:
          type: object
          properties:
            plate: { type: string }
            model: { type: string }
            color: { type: string }
...
    Ride:
      type: object
      required: [id, status, pickup, dropoff, estimatedFareIqd, paymentMethod, requestedAt]
      properties:
        id:     { type: string, format: uuid }
        status: { $ref: '#/components/schemas/RideStatus' }
        rider:  { $ref: '#/components/schemas/PublicUser' }
        driver:
          oneOf:
            - $ref: '#/components/schemas/PublicUser'
            - type: 'null'
        pickup:          { $ref: '#/components/schemas/LatLng' }
        pickupAddress:   { type: string, nullable: true }
        dropoff:         { $ref: '#/components/schemas/LatLng' }
        dropoffAddress:  { type: string, nullable: true }
        estimatedFareIqd:   { type: integer, minimum: 0 }
        finalFareIqd:       { type: integer, nullable: true, minimum: 0 }
        commissionIqd:      { type: integer, nullable: true, minimum: 0 }
        estimatedDistanceM: { type: integer, minimum: 0 }
        estimatedDurationS: { type: integer, minimum: 0 }
        actualDistanceM:    { type: integer, nullable: true, minimum: 0 }
        paymentMethod:      { type: string, enum: [CASH, GATEWAY] }
        requestedAt:        { type: string, format: date-time }
        acceptedAt:         { type: string, format: date-time, nullable: true }
        driverArrivedAt:    { type: string, format: date-time, nullable: true }
        startedAt:          { type: string, format: date-time, nullable: true }
        completedAt:        { type: string, format: date-time, nullable: true }
        cancelledAt:        { type: string, format: date-time, nullable: true }
        cancellationReason: { type: string, nullable: true }
...
    RideOffer:
      type: object
      required: [offerId, rideId, pickup, dropoff, estimatedFareIqd, expiresAt, distanceM]
      properties:
        offerId:  { type: string, format: uuid }
        rideId:   { type: string, format: uuid }
        pickup:   { $ref: '#/components/schemas/LatLng' }
        pickupAddress:  { type: string, nullable: true }
        dropoff:  { $ref: '#/components/schemas/LatLng' }
        dropoffAddress: { type: string, nullable: true }
        estimatedFareIqd: { type: integer, minimum: 0 }
        distanceM: { type: integer, minimum: 0, description: Driver-to-pickup distance. }
        expiresAt: { type: string, format: date-time }
    RideSettlement:
      type: object
      required: [ride, payment, ledgerTransactionId]
      properties:
        ride:    { $ref: '#/components/schemas/Ride' }
        payment: { $ref: '#/components/schemas/Payment' }
        ledgerTransactionId: { type: string, format: uuid }
...
    Wallet:
      type: object
      required: [driverId, balanceIqd]
      properties:
        driverId:   { type: string, format: uuid }
        balanceIqd:
          type: integer
          description: |
            Whole IQD. Derived as SUM(credits) - SUM(debits) over the ledger
            (CLAUDE.md §6.4). May be negative when a driver owes commission.
    SubscriptionPlan:
      type: object
      required: [code, nameAr, nameEn, priceIqd, durationDays]
      properties:
        code:         { type: string, maxLength: 64 }
        nameAr:       { type: string }
        nameEn:       { type: string }
        priceIqd:
          type: integer
          minimum: 0
          description: Whole IQD (CLAUDE.md §6.1). Never a decimal type.
        durationDays: { type: integer, minimum: 1 }
...
    DriverSubscriptionOrNull:
      nullable: true
      allOf:
        - $ref: '#/components/schemas/DriverSubscription'
...
    LedgerEntryPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/LedgerEntry' }
        nextCursor: { type: string, nullable: true }
...